#758: Add LockIsolated file
